Potential demo
Just fooling around with the data available in the sample csv i was
using, I was able to make a dummy algorithm using wind speeds and was
able to output a reasonable power output. So this is something we could
potentially show on monday however this is by no means final.