Final power prediction demo added
I whipped up a final version of the demo (final_ppd.py) which has the
code organized a bit better and with data that outputs reasonable
values. I also tried to explain in comments what is going on. This
algorithm is by no means accurate but it is something and we can explain
that we can implement any other algorithms just as easily for more
accurate predictions.