Updated code for power prediction demo with csv and json
Made two separate scripts for running on csv (the original demo) and a
new one with modifications to accomodate json support. I used the sample
json that I used in the json test script, so I can test later on with a
full data set once we can degrib to json files.