Broadcast variable testing
I tried a couple tests with the broadcast variable for spark. This is
just some example code that I found online. I am also trying to test a
way to connect a spark master and spark slave on my local computer and
on a centos vm with spark installed to further test using a cluster and
would be good for trying to broadcast a variable to multiple nodes.